// server/src/enums/companyApplication.enums.ts
export enum COMPANY_APPLICATION_STATUS {
  PENDING = "pending",
  APPROVED = "approved",
  REJECTED = "rejected",
}

export enum INDUSTRY {
  TECH = "tech",
  FINANCE = "finance",
  HEALTHCARE = "healthcare",
  EDUCATION = "education",
  RETAIL = "retail",
  REAL_ESTATE = "real_estate",
  MANUFACTURING = "manufacturing",
  MEDIA = "media",
  OTHER = "other",
}

export enum COMPANY_SIZE {
  MICRO = "1-10",
  SMALL = "11-50",
  MEDIUM = "51-200",
  LARGE = "201-500",
  ENTERPRISE = "500+",
}