// server/src/enums/company.enums.ts
export enum COMPANY_STATUS {
  PENDING_ACTIVATION = "pending_activation",
  ACTIVE = "active",
  SUSPENDED = "suspended",
}

export enum COMPANY_BENEFIT {
  HEALTH_INSURANCE = "health_insurance",
  SOCIAL_INSURANCE = "social_insurance",
  REMOTE_WORK = "remote_work",
  FLEXIBLE_HOURS = "flexible_hours",
  ANNUAL_BONUS = "annual_bonus",
  PROFIT_SHARING = "profit_sharing",
  TRANSPORTATION_ALLOWANCE = "transportation_allowance",
  MEAL_ALLOWANCE = "meal_allowance",
  TRAINING_AND_DEVELOPMENT = "training_and_development",
  GYM_MEMBERSHIP = "gym_membership",
  CHILDCARE_SUPPORT = "childcare_support",
  WORK_FROM_HOME = "work_from_home",
  OPEN_VACATION_POLICY = "open_vacation_policy",
}